+++ lib/Frontend/pages/intro.dart
@@ -9,6 +9,7 @@ import '../../Backend/logging_wrappers.dart';
 import '../../Backend/plausible_dio.dart';
 import '../../constants.dart';
 import '../../gen/assets.gen.dart';
+import '../Widgets/language_picker.dart';
 import '../Widgets/lottie_lazy_load.dart';
 import '../go_router_config.dart';
 import '../translation_string_definitions.dart';
@@ -32,7 +33,8 @@ class OnBoardingPageState extends ConsumerState<OnBoardingPage> {
     // Navigator.of(context).pushReplacement()
     plausible.event(name: "Complete Onboarding");
     _introLogger.info("Complete Onboarding");
-    HiveProxy.put(settings, hasCompletedOnboarding, hasCompletedOnboardingVersionToAgree);
+    HiveProxy.put(
+        settings, hasCompletedOnboarding, hasCompletedOnboardingVersionToAgree);
     const ActionPageRoute().pushReplacement(context);
   }
 
@@ -46,8 +48,10 @@ class OnBoardingPageState extends ConsumerState<OnBoardingPage> {
   @override
   Widget build(BuildContext context) {
     setupSystemColor(context);
+    ref.watch(initLocaleProvider);
     var pageDecoration = PageDecoration(
-      titleTextStyle: const TextStyle(fontSize: 28.0, fontWeight: FontWeight.w700),
+      titleTextStyle:
+          const TextStyle(fontSize: 28.0, fontWeight: FontWeight.w700),
       bodyTextStyle: const TextStyle(fontSize: 19.0),
       bodyPadding: const EdgeInsets.fromLTRB(16.0, 0.0, 16.0, 16.0),
       pageColor: Theme.of(context).canvasColor,
@@ -92,12 +96,18 @@ class OnBoardingPageState extends ConsumerState<OnBoardingPage> {
           body: homeWelcomeMessage(),
           image: Builder(
             builder: (context) {
-              if (Theme.of(context).colorScheme.brightness == Brightness.light) {
-                return _buildImage(Assets.splashLightTransparent.path, MediaQuery.of(context).size.width);
+              if (Theme.of(context).colorScheme.brightness ==
+                  Brightness.light) {
+                return _buildImage(Assets.splashLightTransparent.path,
+                    MediaQuery.of(context).size.width);
               } else {
-                return _buildImage(Assets.splashDarkTransparent.path, MediaQuery.of(context).size.width);
+                return _buildImage(Assets.splashDarkTransparent.path,
+                    MediaQuery.of(context).size.width);
               }
             },
+          ),
+          footer: LanguagePicker(
+            isButton: true,
           ),
           decoration: pageDecoration,
         ),
@@ -211,7 +221,8 @@ class OnBoardingPageState extends ConsumerState<OnBoardingPage> {
         onPressed: () {
           _onIntroEnd(context);
         },
-        child: Text(onboardingDoneButtonLabel(), style: const TextStyle(fontWeight: FontWeight.w600)),
+        child: Text(onboardingDoneButtonLabel(),
+            style: const TextStyle(fontWeight: FontWeight.w600)),
       ),
       dotsFlex: 1,
       controlsPadding: const EdgeInsets.symmetric(vertical: 32),
@@ -221,7 +232,8 @@ class OnBoardingPageState extends ConsumerState<OnBoardingPage> {
         activeColor: Theme.of(context).colorScheme.primary,
         color: Theme.of(context).colorScheme.tertiary,
         spacing: const EdgeInsets.symmetric(horizontal: 3.0),
-        activeShape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(50.0)),
+        activeShape:
+            RoundedRectangleBorder(borderRadius: BorderRadius.circular(50.0)),
       ),
     );
   }