+++ lib/Frontend/utils.dart
@@ -3,11 +3,13 @@ import 'package:dio/dio.dart';
 import 'package:dio_smart_retry/dio_smart_retry.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:logarte/logarte.dart';
 import 'package:logging/logging.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:platform/platform.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:sentry_dio/sentry_dio.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:wordpress_client/wordpress_client.dart';
@@ -18,46 +20,62 @@ import '../constants.dart';
 import '../l10n/app_localizations.dart';
 import '../l10n/messages_all_locales.dart';
 
+part 'utils.g.dart';
+
 LocalPlatform platform = const LocalPlatform();
 
 Future<bool> getBluetoothPermission(Logger logger) async {
   bool granted = false;
-  if (platform.isAndroid && (await DeviceInfoPlugin().androidInfo).version.sdkInt > 30) {
-    PermissionStatus permissionStatusScan = await Permission.bluetoothScan.request();
+  if (platform.isAndroid &&
+      (await DeviceInfoPlugin().androidInfo).version.sdkInt > 30) {
+    PermissionStatus permissionStatusScan =
+        await Permission.bluetoothScan.request();
     logger.info("permissionStatusScan $permissionStatusScan");
     granted = PermissionStatus.granted == permissionStatusScan;
-    PermissionStatus permissionStatusConnect = await Permission.bluetoothConnect.request();
+    PermissionStatus permissionStatusConnect =
+        await Permission.bluetoothConnect.request();
     logger.info("permissionStatusConnect $permissionStatusConnect");
     granted = granted && PermissionStatus.granted == permissionStatusConnect;
   } else if (platform.isAndroid) {
-    PermissionStatus permissionStatusLocation = await Permission.location.request();
+    PermissionStatus permissionStatusLocation =
+        await Permission.location.request();
     logger.info("permissionStatusLocation $permissionStatusLocation");
     granted = PermissionStatus.granted == permissionStatusLocation;
-    PermissionStatus permissionStatusLocationInUse = await Permission.locationWhenInUse.request();
+    PermissionStatus permissionStatusLocationInUse =
+        await Permission.locationWhenInUse.request();
     logger.info("permissionStatusLocationInUse $permissionStatusLocationInUse");
-    granted = granted && PermissionStatus.granted == permissionStatusLocationInUse;
+    granted =
+        granted && PermissionStatus.granted == permissionStatusLocationInUse;
   } else {
-    PermissionStatus permissionStatusBluetooth = await Permission.bluetooth.request();
+    PermissionStatus permissionStatusBluetooth =
+        await Permission.bluetooth.request();
     logger.info("permissionStatusBluetooth $permissionStatusBluetooth");
     granted = PermissionStatus.granted == permissionStatusBluetooth;
   }
   return granted;
 }
 
-Future<void> initLocale() async {
-  final String defaultLocale = platform.localeName; // Returns locale string in the form 'en_US'
+@Riverpod(keepAlive: true)
+Future<String> initLocale(Ref ref) async {
+  final String defaultLocale =
+      platform.localeName; // Returns locale string in the form 'en_US'
 
-  String? locale = AppLocalizations.supportedLocales
-      .where(
-        (element) => element.toLanguageTag() == HiveProxy.getOrDefault(settings, selectedLocale, defaultValue: ""),
-      )
-      .map(
-        (e) => e.toLanguageTag(),
-      )
-      .firstOrNull;
+  String locale = AppLocalizations.supportedLocales
+          .where(
+            (element) =>
+                element.toLanguageTag() ==
+                HiveProxy.getOrDefault(settings, selectedLocale,
+                    defaultValue: ""),
+          )
+          .map(
+            (e) => e.toLanguageTag(),
+          )
+          .firstOrNull ??
+      defaultLocale;
 
-  await initializeMessages(locale ?? defaultLocale);
-  Intl.defaultLocale = locale ?? defaultLocale;
+  await initializeMessages(locale);
+  Intl.defaultLocale = locale;
+  return locale;
 }
 
 final dioLogger = Logger('Dio');
@@ -116,7 +134,9 @@ Future<WordpressClient> getWordpressClient() async {
   if (_wordpressClient != null) {
     return _wordpressClient!;
   }
-  return WordpressClient.fromDioInstance(baseUrl: Uri.parse('https://thetailcompany.com/wp-json/wp/v2'), instance: await initDio());
+  return WordpressClient.fromDioInstance(
+      baseUrl: Uri.parse('https://thetailcompany.com/wp-json/wp/v2'),
+      instance: await initDio());
 }
 
 Version getVersionSemVer(String input) {
@@ -130,15 +150,20 @@ Version getVersionSemVer(String input) {
   if (split.length > 1 && int.tryParse(split[1]) != null) {
     minor = split[1];
   }
-  if (split.length > 2 && int.tryParse(split[2].replaceAll(RegExp('[^0-9]'), '')) != null) {
+  if (split.length > 2 &&
+      int.tryParse(split[2].replaceAll(RegExp('[^0-9]'), '')) != null) {
     patch = split[2].replaceAll(RegExp('[^0-9]'), '');
   }
-  return Version(major: int.parse(major), minor: int.parse(minor), patch: int.parse(patch));
+  return Version(
+      major: int.parse(major),
+      minor: int.parse(minor),
+      patch: int.parse(patch));
 }
 
 Color getTextColor(Color color) {
   // Counting the perceptive luminance - human eye favors green color...
-  double luminance = (0.299 * color.red + 0.587 * color.green + 0.114 * color.blue) / 255;
+  double luminance =
+      (0.299 * color.red + 0.587 * color.green + 0.114 * color.blue) / 255;
 
   if (luminance > 0.7) {
     return Typography.material2021().black.labelLarge!.color!;
@@ -154,7 +179,8 @@ Future<void> setupSystemColor(BuildContext context) async {
     statusBarIconBrightness: Brightness.dark /*Android=23*/,
     systemStatusBarContrastEnforced: false /*Android=29*/,
     systemNavigationBarColor: Colors.transparent /*Android=27*/,
-    systemNavigationBarDividerColor: Colors.transparent.withAlpha(1) /*Android=28,不能用全透明 */,
+    systemNavigationBarDividerColor:
+        Colors.transparent.withAlpha(1) /*Android=28,不能用全透明 */,
     systemNavigationBarIconBrightness: Brightness.dark /*Android=27*/,
     systemNavigationBarContrastEnforced: false /*Android=29*/,
   );
@@ -166,7 +192,8 @@ Future<void> setupSystemColor(BuildContext context) async {
     systemNavigationBarColor: Colors.transparent,
     // 27
     systemStatusBarContrastEnforced: false /*Android=29*/,
-    systemNavigationBarDividerColor: Colors.transparent.withAlpha(1) /* 不能用全透明 */,
+    systemNavigationBarDividerColor:
+        Colors.transparent.withAlpha(1) /* 不能用全透明 */,
     // 28
     systemNavigationBarIconBrightness: Brightness.dark,
     // 27